requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.20.2
bcrypt==4.2.0
python-dotenv==1.0.1
Pillow==10.4.0
email-validator==2.2.0
python-dateutil==2.9.0
Werkzeug==3.0.3
PyMySQL==1.1.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
